Session light refactoring & audit aggregate root change wip & move things
