+++ Learning.NETCore/AggregateRoot/Learning.AggregateRoot.Domain/Audit/AuditEvent.cs
@@ -1,15 +1,15 @@
 ﻿using System;
 
-namespace Learning.AggregateRoot.Domain
+namespace Learning.AggregateRoot.Domain.Audit
 {
     /// <summary>
-    /// Permet d'auditer toutes les commandes qui ont été réalisées à votre base de données.
+    /// Permet d'auditer tout les évènements qui ont été réalisées à votre base de données.
     /// </summary>
-    public class AuditCommand
+    public class AuditEvent
     {
         public Guid Id { get; set; }
-        public string CommandName { get; set; }
-        public string Command { get; set; }
+        public string EventName { get; set; }
+        public string Event { get; set; }
         public Guid CorrelationId { get; set; }
         public DateTime Date { get; set; }
         public Guid UserId { get; set; }
