+++ Learning.NETCore/AggregateRoot/Learning.AggregateRoot.Infrastructure/Audit/EventAuditer.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Learning.AggregateRoot.Domain;
+using Learning.AggregateRoot.Domain.Audit;
 using Learning.AggregateRoot.Domain.Interfaces.Audit;
 using Learning.AggregateRoot.Domain.Interfaces.AuthentificationContext;
 using Learning.AggregateRoot.Domain.Interfaces.CQRS;
