+++ Learning.NETCore/AggregateRoot/README.md
@@ -3,5 +3,6 @@ Implémentation brouillon des modèles de conceptions AggregateRoot, Session, Re
 Pour réutiliser ce projet il vous faut :
 - Choisir les options que votre aggregate root doit supporter, à savoir avec ou sans audit, suppression logique, event sourcing.
 - Vous pouvez désactiver l'audit de vos commandes et de vos évènements en modifiant votre Médiateur.
+- Vous pouvez désactiver l'audit de tous les changements réalisées à votre base de données en modifiant votre Session.
 - Si vous avez fait le choix de l'audit, créer une implémentation de IAuthentificationContextUserProvider et l'injecter.
 - Si ce n'est pas le cas vous pourrez alors supprimer requestContext, IAuthentificationContext, et tout ce qui va en suivre.
